Resolves #96
In a past commit, the IoArgoprojWorkflowV1alpha1ScriptTemplate creation
in `hera.task.Task` lost it's `volume_mount` attribute. This lead to
script-based tasks not actually mounting any volumes.

This commit re-adds this attribute and adds a test to make sure
script-based tasks actually do mount their volumes in the future.