Add a script constructor that allows for flexible script construction and add a robust runner mode (#543)
Signed-off-by: Sambhav Kothari <skothari44@bloomberg.net>

This PR introduces a new mode for script called callable mode, which
runs the function using the hera runner.

hera runner is a robust version of the current script mode which can
accomodate large files, imports and calling functions with kwargs. It is
highly compatible/tied to pydantic and makes writing argo commands a
breeze using pydantic + hera.

It automatically serializes/deserializes inputs/outputs using pydantic
and also updates the script callable to only return a task when under an
active hera context.

This mode is only possible if the user source code is installed in the
output image along with hera.


---

After conversations with @flaviuvadan, updated the implementation to
allow for the following -

- Added a new ScriptConstructor class that allows users to define their
own ScriptConstructor that can inspect the callable source and generate
the output source. Provided two default implementations - inline and
runner. The default value is `inline` right now, but there are
placeholders in the code to allow for the default value to be detected
automatically. Additionally, users can provide their own
ScriptConstructor if they wish to and set it as a default.
- Added the ability to set class defaults for any hera class using
global_config. This allows users to introduce default values of hera
classes without needing us to implement it for every single use case.
This also removes any circular dependencies between hera global config
and any hera specific classes.

---------

Signed-off-by: Sambhav Kothari <skothari44@bloomberg.net>