Add missing `source` in callable template `kwargs` (#621)
<!-- Thank you for submitting a PR to Hera! 🚀 -->

**Pull Request Checklist**
- [ ] Fixes #<!--issue number goes here-->
- [ ] Tests added
- [ ] Documentation/examples added
- [ ] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title
<!-- Also remember to sign off commits or the DCO check will fail on
your PR! -->

**Description of PR**
<!-- If not linked to an issue, please describe your changes here -->

Currently, the check for source + with_param fails in some instances
when `with_param` is used because `source` is not part of `kwargs`.
Rather, it is part of the callable script that is used to create a
step/task in the callable context. This means it is possible for
`with_param` to not be caught, and hence not intercepted for
constructing `Parameter`s. This PR fixes that

<!-- Piece of cake! ✨🍰✨ -->

---------

Signed-off-by: Flaviu Vadan <flaviuvadan@gmail.com>