[#636] Add special case handling of cron workflow status in service generation (#637)
<!-- Thank you for submitting a PR to Hera! 🚀 -->

**Pull Request Checklist**
- [x] Fixes #636 
- [ ] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title
<!-- Also remember to sign off commits or the DCO check will fail on
your PR! -->

**Description of PR**
<!-- If not linked to an issue, please describe your changes here -->

Currently, the Argo server does not return cron workflow statuses when a
users' cron workflow does not execute _immediately_ post submission. So,
`None`s are returned for the independent fields of the cron workflow
status. While the status is indeed optional, the fields are not! This PR
adds a special case for handling this in the service. I am not sure
whether this is a feature or a bug upstream in Argo Workflows.

CC: @iameskild I was able to run your example using the fix in this PR!

<!-- Piece of cake! ✨🍰✨ -->

---------

Signed-off-by: Flaviu Vadan <flaviuvadan@gmail.com>
Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>
Co-authored-by: Elliot Gunton <egunton@bloomberg.net>