Revert "[#513] Add source args inference to callable template mixin to support source kwargs in regular Py call" (#658)
Reverts argoproj-labs/hera#647

The source kwarg enhancement needs some more tests/consideration for
edge cases before we release it. This PR is to revert on main, and then
we can move the code to an experimental branch to finalise the
interface. It would also be good to use the "experimental feature" flag
when we do release this on main.

Key edge cases to think about
* the script decorator is not checking regular args, only kwargs, so a
function like `hello(name: str)` won't log a warning or fail when
actually used
* At the call site we can be sure when we're looking at Parameters or
Artifact, but plain old data types are less clear
* Therefore an interface that takes all the parameters as args (not
kwargs) could be better in terms of compatibility and edge cases. One
thing to consider is how to pass parameters/artifacts with the new input
name. The `with_name` seems to duplicate `as_argument` and it wouldn't
be clear how to pass the parameter into the function. For artifacts it's
less clear when using a kwarg as the input artifact isn't actually part
of the function kwargs. The general idea would look like:

```py
@script(inputs=[Artifact(name="artifact_1", path=...)])
def my_func(param_1: str, param_2: int = 0):
    ...


# then under a Steps/DAG context:
my_func(
    [
        prev_step.get_parameter("x"),  # how should we specify name "param_1"? .as_argument?
        prev_step.get_parameter("y"),
        prev_step.get_artifact("a"),  # how should we specify name "artifact_1"?
    ],
    name="step-name",
    when=...)

---------

Signed-off-by: Elliot Gunton <egunton@bloomberg.net>