Common arg inference code for steps/tasks (#661)
**Description of PR**
_Copied from
[comment](https://github.com/argoproj-labs/hera/pull/660#issuecomment-1570430679)_
>>I think things like "{{item.name}}" are set automatically for Task
arguments. Should we expose that to Step as well?

>@elliotgunton can confirm these are set automatically for tasks based
on arguments inference. So, Task invocation does not need the arguments
field when with_param is passed, or with_item, but Step does for now. We
should add the same logic to Step as well:
https://github.com/argoproj-labs/hera/blob/main/src/hera/workflows/_mixins.py#L469
Motivation here is to not confuse users with these differences of
"sometimes I can set a thing on Task but not on Step"

---------

Signed-off-by: Elliot Gunton <egunton@bloomberg.net>