Script annotations artifacts (#718)
**Pull Request Checklist**
- [x] Fixes #717 
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
In Hera we are able to specify `Artifact`s in `inputs`:

```python
@script(inputs=Artifact(name="my-artifact", path="/tmp/file"))
def read_artifact():
    with open("/tmp/file") as a_file:
        print(a_file.read())
```

But by using annotations we can avoid repeating the `path` of the file,
and access the artifact as if it's a regular Python argument of the
given type:

```python
@script()
def read_artifact(an_artifact: Annotated[Path, Artifact(name="my-artifact", path="/tmp/file")]):
    print(an_artifact.read_text())
```

The path duplication is no longer necessary since `an_artifact` is a
`Path` object.

The fields allowed in the `Artifact` annotations are: `name`, `path`,
and `loader`.
For `Artifact`s, we allow three types of loaders. Using
`ArtifactLoader`, we have `file`, `json`,
and `None`.
With no loader, the `path` attribute of `Artifact` is extracted and can
be subsequently used in the function body by referring to the function
parameter. This can be seen above.
When the loader is set to `file`, the function parameter will be the
contents of the file
stored at `path`. 

```python
@script()
def read_artifact(
    an_artifact: Annotated[str, Artifact(name="my-artifact", path=ARTIFACT_PATH, loader=ArtifactLoader.file)]
) -> str:
    return an_artifact
```

This loads the contents of the file at `ARTIFACT_PATH` to the argument
`an_artifact` and subsequently can be used as a string inside the
function.

When the loader is set to `json`, the contents of the file at `path` are
read and parsed to `json`.

```python
class MyArtifact(BaseModel):
    a = "a"
    b = "b"


@script()
def read_artifact(
    an_artifact: Annotated[MyArtifact, Artifact(name="my-artifact", path=ARTIFACT_PATH, loader=ArtifactLoader.json)]
) -> str:
    return an_artifact.a + an_artifact.b
```

Here, we have a json representation of `MyArtifact` stored at
`ARTIFACT_PATH`. We can load it with `ArtifactLoader.json` and then use
`an_artifact` as an instance of `MyArtifact` inside the function.

---------

Signed-off-by: Mikolaj Deja <mdeja2@bloomberg.net>
Co-authored-by: Mikolaj Deja <mdeja2@bloomberg.net>