Fix errors with underscores in labels (#772)
**Pull Request Checklist**
- [ ] Fixes #<!--issue number goes here-->
- [ ] Tests added
- [ ] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

When submitting a workflow with the new outputs annotation, this error
occurs:

```
[spec.volumes[2].name: Invalid value: "hera__outputs_directory": a lowercase RFC 1123 label must consist 
of lower case alphanumeric characters or '-', and must start and end with an alphanumeric character 
(e.g. 'my-name', or '123-abc', regex used for validation is '[a-z0-9]([-a-z0-9]*[a-z0-9])?'), 
spec.containers[0].volumeMounts[0].name: Not found: "hera__outputs_directory", 
spec.containers[1].volumeMounts[0].name: Not found: "hera__outputs_directory"]
```

This PR fixes the issue by changing the underscores to dashes in labels:
Volumes and Steps. It also renames env variables to use capital letters.

Signed-off-by: Mikolaj Deja <mdeja2@bloomberg.net>
Co-authored-by: Mikolaj Deja <mdeja2@bloomberg.net>
Co-authored-by: Flaviu Vadan <flaviuvadan@gmail.com>