Add example of how to run computation against a Dask cluster (#805)
**Pull Request Checklist**
- [ ] Fixes #<!--issue number goes here-->
- [ ] Tests added
- [x] Documentation/examples added
- [ ] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
I received feedback that some users would like some suggestions on how
to integrate with external products such as Dask, Spark, etc. This PR
introduces a first such example of using Dask in Hera.

**Question**
Should we have an `integrations` folder to make it clear what this type
of example is about? I think it's ok for now to leave it as a Hera
example but perhaps in the future it can be organized in its own
namespace

Signed-off-by: Flaviu Vadan <flaviuvadan@gmail.com>