Bug fixes for input Parameters and Artifacts in annotations (#811)
**Pull Request Checklist**
- [x] Fixes #809, fixes #810
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
Currently, output parameters can be specified without a name within the
function parameters, but input parameters cannot. This PR makes the
runner use the function parameter name as the `name` of a
`hera.workflows.Parameter`.

Input Artifacts can be specified without a path in the annotation, which
creates the correct yaml, seen in #792, but at runtime the runner would
not know what path the artifact is supposed to load from. Fixed and
tested.

This PR also cleans up the tests directory structure, removing a lot of
duplicated code for runner/annotation tests. The distinction between
these is now made clear in docstrings for `test_runner.py` and
`test_script_annotations.py` - the runner tests should run the actual
script function code to emulate the Argo cluster, while the script
annotation tests check that function definitions produce correct yaml
and are valid Python when using the annotations.

---------

Signed-off-by: Elliot Gunton <egunton@bloomberg.net>