Add Pydantic `Extra.forbid` to shared base model config  (#820)
…examples

**Pull Request Checklist**
- [x] Fixes #819
- [x] Tests exists
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
This PR adds the Pydantic `Extra.forbid` configuration to the shared
base model. This means users will no longer be faced with silent
semantic errors as outlined in
https://github.com/argoproj-labs/hera/issues/819. The core change is the
base model. The rest of the changes are related to fixes after
introducing the new config. For instance, there were Hera examples that
used `subpath` vs `sub_path`, which is exactly the sort of problem users
have!

---------

Signed-off-by: Flaviu Vadan <flaviuvadan@gmail.com>
Signed-off-by: Sambhav Kothari <skothari44@bloomberg.net>
Co-authored-by: Sambhav Kothari <skothari44@bloomberg.net>
Co-authored-by: Elliot Gunton <egunton@bloomberg.net>