Add support for modifying field properties prior to model generation (#874)
**Pull Request Checklist**
- [x] Fixes #869
- [x] Tests added (we already have tests / examples that compare
upstream and Hera YAMLs that include HTTPGetAction)
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

#869 reports how the field `IntOrString` incorrectly renders the port
specification into a string that's not accepted by K8s. This PR adds a
spec modification strategy that takes the remote Argo OpenAPI JSON
specification and modifies the field from `IntOrString` to `integer`.
This way the correct model is generated, and users know an integer is
supposed to be used. Ideally, we'd fix these upstream but this at least
unblocks current Hera users.

---------

Signed-off-by: Flaviu Vadan <flaviuvadan@gmail.com>