Implement get_parameter API on IO mixin and add tests (#876)
**Pull Request Checklist**
- [x] Fixes #816
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

Implements the `get_parameter` API on the `IOMixin`. This allows clients
that need parameters in `arguments` fields to get the parameter object
from DAGs, Steps, etc. This helps avoid the need to write
`{{inputs.parameters.whatever}}` explicitly. `get_parameter` _assumes_
that clients want to use a parameter as input, so the value field is set
accordingly

Signed-off-by: Flaviu Vadan <flaviuvadan@gmail.com>