Update the "bug occurs when" section
There are 3 main categories I can see for bug reports, exporting to YAML, submitting to Argo (which is actually likely related to YAML/`to_dict` exporting), and running on Argo with the Hera runner, now that it is graduating from experimental. Added an "Other" option as well for reporter to explain.

Signed-off-by: Elliot Gunton <egunton@bloomberg.net>