Add example with Weights and Biases + distributed finetuning (#932)
I've been meaning to add these examples to the repo for a while. One is
an example of how one can use WandB with Hera scripts, which is really
not very different than the core Python example already documented
[here](https://wandb.ai/home). The other example is from the talk I had
the chance to give with @JPZ13 at KubeCon NA 2023. The example added in
this PR is a flat representation of [this
repo](https://github.com/flaviuvadan/kubecon_na_23_llama2_finetune). My
hope with the two is illustrate that distributed training is much more
accessible in Hera + Argo Workflows on K8s. Some useful references:
- https://openai.com/research/scaling-kubernetes-to-7500-nodes
- https://kubernetes.io/case-studies/openai/

---------

Signed-off-by: Flaviu Vadan <flaviuvadan@gmail.com>
Co-authored-by: Elliot Gunton <egunton@bloomberg.net>