Move Hera's copyright from Dyno Therapeutics to "The Hera Authors" (#957)
We've had multiple discussions about this topic in the past few months.
We already took the step to move from MIT to Apache 2.0 and we needed to
switch the copyright as well. Dyno Therapeutics reviewed this change and
approved it internally. This PR changes the copyright from `2021 Dyno
Therapeutics Inc` to `2024 The Hera Authors`.

Signed-off-by: Flaviu Vadan <flaviuvadan@gmail.com>