Fix RunnerInput defaults (#974)
**Pull Request Checklist**
- [x] Fixes #962 
- [x] Tests added
- [ ] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
Currently, defaults are required when using Parameters in the
RunnerInput. This PR allows defaults to be omitted.

## Changes from #977 
- [x] Fixes #962 
- [x] Tests added
- [ ] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR title

**Description of PR**
Primarily, this PR refactors the runner code. It also fixes the mapping and loading of kwargs to a RunnerInput object.

Currently, the runner code is hard to follow. This PR refactors the functionality in the `runner.py` file into a `_runner` module with util submodules. The PR also makes the complex input mapping logic more testable.

## Changes from #982 

* Script user guide was too long, split into main features. Fix internal links
* Make pydantic io example into a runnable workflow - made it more obvious the scripts would need a custom image

**Pull Request Checklist**
- [x] Fixes #961 (fixes example)
- [ ] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR title

**Description of PR**
Currently, the pydantic IO example does not work. Combined with fixes in PRs #974 and #977, this doc change shows how to use the Runner IO features.

---------

Signed-off-by: Elliot Gunton <egunton@bloomberg.net>