Fix `to` generating to folder and refactor tests (#989)
* Previously, if a user specifies a `to` option such as `foo.yaml`, the
cli will create the target as a folder `foo.yaml/`. This change inspects
the suffix of the `to` path; if it is a recognised "yaml" extension, we
create the file's parents, and writes the workflow to the file. We
therefore do not need to check if the target exists
* The commit also refactors some code for testability and simplifies the
tests themselves for readability to avoid magic number accesses

---------

Signed-off-by: Elliot Gunton <egunton@bloomberg.net>