Add SyntaxError for declaring outside of a Workflow context (#1104)
**Pull Request Checklist**
- [x] Fixes #921 
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title


Changes `assert` added previously into a full SyntaxError for better
diagnostics for users. We could allow declaration of `Steps`/`Tasks`
outside to then add later but it would overcomplicate the build process
(to loop through the nodes and add the templates). Plus this error
message should not be possible when users move to the new decorators
syntax.

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>