Refactor init_containers building logic in TemplateMixin (#1157)
**Pull Request Checklist**
- [x] Fixes #1155 
- [ ] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**PR Description**

When using Hera to generate an Argo Workflow, I encountered an issue
with the env structure in initContainers. Specifically, the env variable
references for secrets are not being properly constructed. Instead of
the correct structure with valueFrom, the generated YAML output directly
assigns secret_name and secret_key, which is not valid in Argo
Workflows.

This issue occurs because Hera doesn't handle this internally within its
template constructors. Instead, it uses the `initContainers` as they are
provided
[here](https://github.com/argoproj-labs/hera/blob/main/src/hera/workflows/container.py#L87).
To address this, I created the `_build_init_containers` method in the
`TemplateMixin` class.

```python
    def _build_init_containers(self) -> Optional[List[ModelUserContainer]]:
        """Builds the `init_containers` field and optionally returns a list of `UserContainer`."""
        if self.init_containers is None:
            return None

        return [i.build() if isinstance(i, UserContainer) else i for i in self.init_containers]
```
This method ensures that the `init_containers` field is correctly
constructed, and if any of the containers are instances of
`UserContainer`, they are properly built before being returned.

Signed-off-by: sh.yoon <sh.yoon.devops@gmail.com>
Co-authored-by: Sambhav Kothari <skothari44@bloomberg.net>