Use overloads in script decorator types
Returning a union of callables from @script-decorated user functions makes it impossible to type-check without suppressing mypy's 'call-arg' error (or equivalent), which completely disables parameter type-checking. Instead, use an overloaded callable Protocol to select the signature most likely to be what the user was intending to pick.

The call signatures to Step and Task have to be duplicated, as:
 - mypy and pyright do not support having multiple ParamSpec-decorated overloads
 - the `name` parameter is required in Step and Task but optional when calling a @script-decorated user function

Signed-off-by: Alice Purcell <alicederyn@gmail.com>