Factor out common workflow annotation iteration
Multiple methods in _io_mixins resolve the same problem of iterating
through all fields, constructing a workflow annotation based on the
user's annotations, defaulting the name if unset or falling back to a
Parameter if no annotation is found. Factor that logic out into an
iterator, simplifying the various methods and reducing the risk of
discrepancies and bugs in future.

Signed-off-by: Alice Purcell <alicederyn@gmail.com>