Disable warnings from model_dump (#1214)
**Description of PR**
Currently, using Pydantic V2 with the Input/Output mixin types and a
non-`str` field prints a `UserWarning` when the class is declared, of
the form "UserWarning: Pydantic serializer warning: Expected `int` but
got `str` with value `'{{...}}'` - serialized value may not be as
expected". This is caused by using `model_dump` on an object constructed
with all string values, i.e. one which does not pass the type-checking
performed by the method, as part of class creation.

This PR just adds `warnings="none"` to the two uses of `model_dump`, to
disable these warnings.

Signed-off-by: Alice Purcell <alicederyn@gmail.com>