Fix TypeError in script.py for functions returning non-types (#1224)
**Pull Request Checklist**
- [ ] Fixes #<!--issue number goes here-->
- [X] Tests added
- [ ] Documentation/examples added
- [X] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
Currently, `_get_outputs_from_return_annotation` is always passing the
return type annotation (if one exists) to `issubclass`, to check if it
is a Pydantic type. Some valid type annotations like `typing.Dict` are
not valid types at runtime, i.e. cannot be used with `subclass`,
resulting in a runtime `TypeError`.

This PR changes the precondition for entering this code from truthiness
to `isinstance(..., type)`, matching the code a few lines above.

Signed-off-by: Alice Purcell <alicederyn@gmail.com>