Extract construct_io_from_annotation function
Many places need to inspect a parameter or field to construct a
Parameter or Artifact based on the name and other annotations. Extract
the single-field logic from _construct_io_from_fields to simplify many
callsites that were using get_workflow_annotation and stitching in the
name and default Parameter themselves. This will also allow us to
reliably provide more features in future, such as inferring an enum
parameter for a Literal.

Signed-off-by: Alice Purcell <alicederyn@gmail.com>