Fix runtime error in origin_type_is*type functions
If origin_type_issubtype or origin_type_issupertype are passed a special
form annotation, they will raise a TypeError due to passing it into
issubclass. Fix this issue by first checking if the annotation is a
type, and returning False if not.

Unit test this with NoReturn, as we are very unlikely to ever create
a special case for this type in the function.

Signed-off-by: Alice Purcell <alicederyn@gmail.com>