Fix compatibility with Argo Workflows 3.6 (#1293)
**Pull Request Checklist**
- [x] Fixes #1215
- [x] Fixes #1292 
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
Updates auto-generated models to Argo 3.6. Note there are a few changes
to Hera code which are technically breaking changes (but are very simple
for users to change):
* `resources: Optional[ResourceRequirements]` is now `resources:
Optional[VolumeResourceRequirements]` for volumes. The attributes in the
classes are the same except for `claims` is removed in
`VolumeResourceRequirements`. Updated 4 examples that were using this
* `cluster_name` has been removed in `ObjectMeta` so is also removed in
the `Workflow` class. It was never used by the Argo server (and is
explicitly noted as an "ignored" field in the field description in
`ObjectMeta`).

The rest are changes in the Argo models (i.e. in
`src/hera/workflows/models/io/argoproj/workflow/v1alpha1.py`), that will
affect users using the models directly. **i.e. They'll be
(relatively<sup>1</sup>) stuck on the current `Hera` version until they
upgrade Argo to 3.6 if they are using the following classes in
Argo<3.6**:
* `PodGC` structurally changed - see #1292
* Multiple classes removed with the `Artifact` prefix
* ArtifactGCSpec, ArtifactGCStatus, ArtifactNodeSpec, ArtifactResult,
ArtifactResultNodeStatus
* Duration and Effect removed (now just `str`)
* Container imagePullPolicy is now a str not an enum (we handled either
case in `ContainerMixin` so no change on the Hera side)

1. The workaround would be to use `ModelClass.construct` as seen
[here](https://github.com/argoproj-labs/hera/issues/1292#issuecomment-2536592975).

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>