Add separate make targets for upstream examples
* If the upstream examples are changed while working on a feature, it
can be frustrating to regenerate YAML files as they cannot match without
manually editing the Python versions, which mixes up intentions in the PR.
* For that reason, use separate make targets for regenerating the Hera-generated YAML files
(`regenerate-upstream-test-data`) and the actual upstream YAML files
(`fetch-upstream-examples`), so move the test_for_missing_examples step.
* Add a check to ensure that the local copy of the upstream YAML matches
for `regenerate-upstream-test-data`, with a test fail message to tell
user to fetch the new YAML with the `fetch-upstream-examples` target.
* Update the `regenerate-test-data` target to only regenerate non-upstream
examples

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>