Normalise arguments on assignment (#1325)
**Pull Request Checklist**
- [x] Fixes #1169 
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
When reassigning the `arguments` field of a workflow (e.g. to submit the
same workflow multiple times with different arguments), the validator
should be reapplied to ensure that the field is a list. This is because
the field can be assigned to as a dictionary which causes an error when
building the workflow. This is a special case just for the arguments
field, no plans to do the same with inputs/outputs/etc.

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>