Fix `with_items` handling for scripts (#1332)
**Pull Request Checklist**
- [x] Fixes #1314 
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
* We were always assuming a list of dicts in `with_items` when used with
a script template
* Now check if `with_items` is a list of dicts, and perform validation
to ensure all dicts have the same keys that match the source parameter
names, and prompt user to serialize if not. Also check that all (or none) of
the items are dicts.
* If `with_items` is a list of plain values, we now return `None` as
there are no parameter substitutions to be made
* Renamed functions for clarity and improved docstrings (this section of
the code has always confused me)
* Add an example for different usage of `with_items` in a script
template

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>