Create new docs homepage (#1343)
**Pull Request Checklist**
- [x] Fixes #1318 
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
The docs homepage was a symlinked copy of the repo README, which isn't
necessarily appropriate for the docs. This PR adds a separate `index.md`
for the docs homepage, with some nice styling, and some motivation of
what Hera is and why use it:


![image](https://github.com/user-attachments/assets/f1d9e90e-08d3-46fb-b17a-c00b2f882c65)

Also streamlined the repo README to remove some old information, show
only 1 example, and collapse the list of presentations. I also fixed
some typos seen while editing (in the third commit).

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>