Update deps and examples (#1363)
**Description of PR**
Currently, there are a load of deps to be updated, including mypy which
raised a complaint about the missing `**kwargs` in the new decorators
(which is by design). I wanted to avoid polluting the git history with
all the dependabot PRs so have closed them.

This PR updates deps, regenerates docs, fixes test/linting complaints.

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>