Add Best Practices user guide (#1401)
**Pull Request Checklist**
- [x] Fixes #1400 (but is in no way final)
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
Add an initial set of best practices to create a new user guide. This
isn't meant to be the "final" version as it should be more of a living
document that we'll iterate on over time.

As part of this I also created
https://github.com/elliotgunton/hera-example-project which can be used
to exemplify the best practices and design patterns, especially around
Runner scripts.

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>