Preserve folder structure when using CLI code conversion (#1443)
**Pull Request Checklist**
- [x] Fixes comments raised in #1440 
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
This is mostly a refactor of the CLI code with a small feature added.
The `--recursive` flag now by default preserves the input directory
structure when outputting to a folder, i.e. they will have the same
structure. The `--flatten` flag was added to allow the old behaviour,
which is still useful for outputting many files to a single, flat
folder.

In terms of refactors:
* Extract file interactions into `util.py`
* Use `Path` functions directly where possible
* Simplify test_generate_yaml to use `tmp_path` instead of `os` mock_os
* Reword CLI help
* Extract conversion logic to a new `convert_code` function in `util.py`

Also added all the equivalent YAML-to-Python CLI tests as the
(simplified) test_generate_yaml tests.

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>