Fix module for script runners generated by CLI and from `__main__` (#1445)
**Pull Request Checklist**
- [x] Fixes #573, Fixes #1161
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
Currently, Script Runners cannot be exported to YAML correctly if they
are contained in the same module as `__main__`, i.e. when trying to
export a workflow using the file it's written in, as the
`transform_values` sees `__main__` as the module. The CLI also ignores
the full path spec of the function, only using the stem.

This PR fixes both cases by using a utility function to construct a
valid module path.

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>