Add Hera `RetryStrategy` to `TemplateMixin` and `Workflow` (#1455)
**Pull Request Checklist**
- [x] Fixes #1454 
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
Currently, the `RetryStrategy` from `hera.workflows.retry_strategy` is
not used for fields of `TemplateMixin` and `Workflow`, which actually
use the `RetryStrategy` from `hera.workflows.models`, meaning the use of
the Hera class will fail linting (but otherwise worked correctly as the
fields were the same). This PR updates the type to the a union of the
two, and adds the build function.

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>