Update on-cluster testing (#1466)
**Pull Request Checklist**
* Fixes #1397 
* Fixes #1350 
* Fixes #1163 

- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
This PR adds an improved "on cluster" test harness, which will run all
non-upstream examples.
* CI will not run all the non-upstream examples, but will continue to
run the 3 basic examples.
* CI will also `lint` all the non-upstream examples (which runs in
around 1 second on local).

Added `pytest-xdist` specifically for this test otherwise there would be
a lot of sitting around waiting. I updated examples which could be
easily fixed, and created a list of examples which should be skipped for
linting and submission (the submission set subsumes the linting set).
(Only ~one~ two examples are intentionally skipped for linting!)

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>