Fix incorrect types and comments throughout mixins (#1472)
Was intending to look into #1254, but found some odd types that looked
wrong. This PR is mainly a refactor to fix types, with a minimal
addition of the `raise ValueError("Only `template` is supported (not
inline or template_ref)")` snippets (as that would be the correct fix
which was unearthed through these type fixes).

To note:
* TTemplate had the wrong description and so was used incorrectly in
places where the `isinstance` check would never be true
* Prefer `Templatable` as a type over `TemplateMixin`

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>