Add example using `with_items` for step/dag decorators (#1490)
Received a question on how to perform fan outs using the decorator
syntax. It is already possible, so I added an example and expanded the
user guide. Added multiple warnins as to what's keeping decorator syntax as
"experimental", as I've received a few questions around that too -
essentially we just want more feedback and to improve the UX as much as
possible before graduating the feature.

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>