Add optional artifact support in Hera runner (#1497)
**Pull Request Checklist**
- [x] Fixes #1253 
- [x] Tests added
- [x] Documentation/examples added
- [x] [Good commit messages](https://cbea.ms/git-commit/) and/or PR
title

**Description of PR**
This PR neatly fixes the bug described in #1253 with `if
artifact_annotation.optional and not path.exists():`. This has been
tested on a local install of Argo Workflows.

This uncovered the problem that if an `optional=True` Artifact is
actually `None` at runtime, the type must be `Optional`:
```
pydantic_core._pydantic_core.ValidationError: 1 validation error for artifact_loaders
a_file_as_json_optional
  Input should be a valid dictionary [type=dict_type, input_value=None, input_type=NoneType]
    For further information visit https://errors.pydantic.dev/2.11/v/dict_type
```

Therefore, this PR also adds enhancements so that:

1. the type annotation must be `Optional` if the Artifact is optional
2. if the type is `Optional` then users don't actually have to set
`optional=True`

I have also added validation to ensure users do not mix
optional/non-optional manually, as this could create runtime validation
errors later if the Artifact does not exist (when type is not `Optional`
but it is `None` at runtime).

---------

Signed-off-by: Elliot Gunton <elliotgunton@gmail.com>